+++ trixnity-client/src/commonMain/kotlin/de/connect2x/trixnity/client/crypto/ClientOlmStore.kt
@@ -1,5 +1,6 @@
-package de.connect2x.trixnity.client.cryptodriver
+package de.connect2x.trixnity.client.crypto
 
+import de.connect2x.lognity.api.logger.Logger
 import de.connect2x.trixnity.client.key.getDeviceKey
 import de.connect2x.trixnity.client.store.AccountStore
 import de.connect2x.trixnity.client.store.KeyStore
@@ -10,6 +11,7 @@ import de.connect2x.trixnity.client.store.getByStateKey
 import de.connect2x.trixnity.client.store.members
 import de.connect2x.trixnity.client.store.toDeviceTrustLevel
 import de.connect2x.trixnity.client.user.LoadMembersService
+import de.connect2x.trixnity.core.UserInfo
 import de.connect2x.trixnity.core.model.RoomId
 import de.connect2x.trixnity.core.model.UserId
 import de.connect2x.trixnity.core.model.events.m.room.EncryptionEventContent
@@ -19,12 +21,16 @@ import de.connect2x.trixnity.core.model.keys.EncryptionAlgorithm
 import de.connect2x.trixnity.core.model.keys.KeyValue.Curve25519KeyValue
 import de.connect2x.trixnity.core.model.keys.SignedDeviceKeys
 import de.connect2x.trixnity.crypto.key.DeviceTrustLevel
+import de.connect2x.trixnity.crypto.olm.OlmStore
 import de.connect2x.trixnity.crypto.olm.StoredInboundMegolmMessageIndex
 import de.connect2x.trixnity.crypto.olm.StoredInboundMegolmSession
 import de.connect2x.trixnity.crypto.olm.StoredOlmSession
 import de.connect2x.trixnity.crypto.olm.StoredOutboundMegolmSession
+import de.connect2x.trixnity.crypto.sign.SignService
 import kotlin.time.Instant
 import kotlinx.coroutines.flow.first
+
+private val log = Logger("de.connect2x.trixnity.client.cryptodriver")
 
 class ClientOlmStore(
     private val accountStore: AccountStore,
@@ -33,10 +39,20 @@ class ClientOlmStore(
     private val roomStateStore: RoomStateStore,
     private val loadMembersService: LoadMembersService,
     private val tm: StoreTransactionManager,
-) : de.connect2x.trixnity.crypto.olm.OlmStore {
+    private val signService: SignService,
+    private val userInfo: UserInfo,
+) : OlmStore {
 
-    override suspend fun getDeviceKeys(userId: UserId): Map<String, SignedDeviceKeys>? =
-        keyStore.getDeviceKeys(userId).first()?.mapValues { it.value.value }
+    override suspend fun getDeviceKeys(userId: UserId): Map<String, SignedDeviceKeys>? {
+        val deviceKeys = keyStore.getDeviceKeys(userId).first()?.mapValues { it.value.value } ?: return null
+        if (userId != userInfo.userId) return deviceKeys
+        val ownDeviceKeys = deviceKeys[userInfo.deviceId] ?: return deviceKeys
+        if (!ownDeviceKeys.signatures.isNullOrEmpty()) return deviceKeys
+        log.debug {
+            "fallback to manually signed own device keys, because of a historical json bug, signatures may be missing in the database"
+        }
+        return deviceKeys + (userInfo.deviceId to signService.getSelfSignedDeviceKeys())
+    }
 
     override suspend fun getMembers(roomId: RoomId, memberships: Set<Membership>): Set<UserId> {
         loadMembersService(roomId, true)